Page-context init scripts for killers.
Killer may execute some code in page context (which means it can access
and modify javascript objects created by scripts on the page). That
could be used to hijack already existing UI and make it control CTP's
player. To have the init script executed, just pass source code of the
init function as initScript property of mediaData. This should be a
function with one argument - DOM ID of the media element.

Unfortunately, scripts executed in page context don't have access to
MediaPlayer and other objects in CTP context, so the only reliable
communication method is event dispatching. There's currently a very
simple RPC mechanism: using the dataset property of the media element,
set dataset.ctpCommandName to the name of the command you want to
execute, dataset.ctpCommandArgument to JSON-serialized argument and then
dispatch a 'ctpcommand' event on the media element. Right now the only
available command is seekTo(time), but it's very easy to add more.

Note: this function probably should add a handler for 'ctpdestroy' event
on the media element to restore hijacked functionality - this would be
called if user goes back to the Flash player.